backend/requirements.txt
# Core
streamlit>=1.35.0
google-generativeai>=0.8.0
google-genai>=0.5.0
pypdf>=4.0.0
Pillow>=10.0.0
fpdf2>=2.7.0

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-docx>=1.1.0

# Testing
pytest>=8.0.0
